Organize popup page into a dedicated folder
This will move all popup-related files into a separate  directory for better organization.
