--- popup.js
@@ -1,7 +1,0 @@
-import { initializeButtonEvents } from "./buttonEvents.js";
-import { initializeOnLoad } from "./onLoad.js";
-
-$(function () {
-  initializeButtonEvents();
-  initializeOnLoad();
-});
