+++ popup/js/index.js
@@ -1,0 +1,7 @@
+import { initializeButtonEvents } from "./buttonEventHandler.js";
+import { initializeOnLoad } from "./onLoadInitialization.js";
+
+$(function () {
+  initializeButtonEvents();
+  initializeOnLoad();
+});
